+++ src/main/java/net/neoforged/moddevgradle/internal/ModDevPlugin.java
@@ -12,6 +12,11 @@ import net.neoforged.moddevgradle.internal.utils.ExtensionUtils;
 import net.neoforged.moddevgradle.internal.utils.FileUtils;
 import net.neoforged.moddevgradle.internal.utils.IdeDetection;
 import net.neoforged.moddevgradle.internal.utils.StringUtils;
+import net.neoforged.moddevgradle.nfrt.ArtifactManifestEntry;
+import net.neoforged.moddevgradle.nfrt.CreateArtifactManifestTask;
+import net.neoforged.moddevgradle.nfrt.CreateMinecraftArtifactsTask;
+import net.neoforged.moddevgradle.nfrt.DownloadAssetsTask;
+import net.neoforged.moddevgradle.tasks.NeoFormRuntimeTask;
 import net.neoforged.moddevgradle.tasks.JarJar;
 import net.neoforged.vsclc.BatchedLaunchWriter;
 import net.neoforged.vsclc.attribute.ConsoleType;
@@ -76,7 +81,6 @@ import java.util.IdentityHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
-import java.util.Set;
 import java.util.function.Consumer;
 import java.util.function.Function;
 import java.util.stream.Collectors;
