+++ src/main/java/net/neoforged/moddevgradle/tasks/NeoFormRuntimeTask.java
@@ -1,4 +1,4 @@
-package net.neoforged.moddevgradle.internal;
+package net.neoforged.moddevgradle.tasks;
 
 import net.neoforged.moddevgradle.internal.utils.IdeDetection;
 import net.neoforged.moddevgradle.internal.utils.NetworkSettingPassthrough;
